Move large chunks of code into separate modules
