+++ iroh-bytes/src/store/redb.rs
@@ -1,7 +1,7 @@
 //! redb backed storage
 
 use std::{
-    collections::{btree_map, BTreeMap, BTreeSet},
+    collections::{btree_map, BTreeMap},
     fs::OpenOptions,
     io::{self, BufReader, Read, Write},
     path::{Path, PathBuf},
@@ -16,7 +16,6 @@ use bao_tree::io::{
 };
 use bytes::Bytes;
 use futures::{channel::oneshot, Stream, StreamExt};
-use std::str::FromStr;
 
 use iroh_base::hash::{BlobFormat, Hash, HashAndFormat};
 use iroh_io::AsyncSliceReader;
@@ -25,6 +24,11 @@ use serde::{Deserialize, Serialize};
 use smallvec::SmallVec;
 use tokio::io::AsyncWriteExt;
 use tracing::trace_span;
+
+mod import_flat_store;
+#[cfg(test)]
+mod tests;
+mod validate;
 
 use crate::{
     store::bao_file::{BaoFileStorage, CompleteMemOrFileStorage},
@@ -38,7 +42,7 @@ use crate::{
 use super::{
     bao_file::{raw_outboard_size, BaoFileConfig, BaoFileHandle, CreateCb, DropCb},
     temp_name, BaoBatchWriter, EntryStatus, ExportMode, ExportProgressCb, ImportMode,
-    ImportProgress, MapEntry, ReadableStore, TempCounterMap, ValidateLevel, ValidateProgress,
+    ImportProgress, MapEntry, ReadableStore, TempCounterMap, ValidateProgress,
 };
 
 use super::BaoBlobSize;
@@ -707,7 +711,7 @@ pub(crate) enum ActorMessage {
     /// Modification method: marks a partial entry as complete.
     /// Calling this on a complete entry is a no-op.
     OnComplete { hash: Hash },
-    ImportEntry {
+    Import {
         content: HashAndFormat,
         file: ImportSource,
         data_size: u64,
@@ -778,6 +782,7 @@ pub(crate) enum ActorMessage {
     Shutdown,
 }
 
+/// Predicate for filtering entries in a redb table.
 pub(crate) type FilterPredicate<K, V> =
     Box<dyn Fn(u64, AccessGuard<K>, AccessGuard<V>) -> Option<(K, V)> + Send + Sync>;
 
@@ -1267,7 +1272,7 @@ impl StoreInner {
         let hash = *tag.hash();
         // blocking send for the import
         let (tx, rx) = flume::bounded(1);
-        self.tx.send(ActorMessage::ImportEntry {
+        self.tx.send(ActorMessage::Import {
             content: HashAndFormat { hash, format },
             file,
             outboard,
@@ -2153,369 +2158,25 @@ impl Actor {
                     }
                 }
             }
-            tx.commit()?;
+            match tx.commit() {
+                Ok(()) => {
+                    for path in delete_after_commit {
+                        if let Err(cause) = std::fs::remove_file(&path) {
+                            tracing::error!("failed to remove file: {}", cause);
+                        };
+                    }
+                }
+                Err(cause) => {
+                    for path in delete_on_fail {
+                        if let Err(cause) = std::fs::remove_file(&path) {
+                            tracing::error!("failed to remove file: {}", cause);
+                        };
+                    }
+                    return Err(cause.into());
+                }
+            }
         }
         Ok(())
-    }
-
-    fn import_flat_store(&mut self, paths: FlatStorePaths) -> ActorResult<bool> {
-        /// A file name that indicates the purpose of the file.
-        #[derive(Clone, PartialEq, Eq)]
-        pub enum FileName {
-            /// Incomplete data for the hash, with an unique id
-            PartialData(Hash, [u8; 16]),
-            /// File is storing data for the hash
-            Data(Hash),
-            /// File is storing a partial outboard
-            PartialOutboard(Hash, [u8; 16]),
-            /// File is storing an outboard
-            ///
-            /// We can have multiple files with the same outboard, in case the outboard
-            /// does not contain hashes. But we don't store those outboards.
-            Outboard(Hash),
-            #[allow(dead_code)]
-            /// Temporary paths file
-            TempPaths(Hash, [u8; 16]),
-            /// External paths for the hash
-            Paths(Hash),
-            /// File is going to be used to store metadata
-            Meta(Vec<u8>),
-        }
-
-        impl FileName {
-            /// Get the file purpose from a path, handling weird cases
-            pub fn from_path(path: impl AsRef<Path>) -> std::result::Result<Self, &'static str> {
-                let path = path.as_ref();
-                let name = path.file_name().ok_or("no file name")?;
-                let name = name.to_str().ok_or("invalid file name")?;
-                let purpose = Self::from_str(name).map_err(|_| "invalid file name")?;
-                Ok(purpose)
-            }
-        }
-
-        impl FromStr for FileName {
-            type Err = ();
-
-            fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-                const OUTBOARD_EXT: &str = "obao4";
-                // split into base and extension
-                let Some((base, ext)) = s.rsplit_once('.') else {
-                    return Err(());
-                };
-                // strip optional leading dot
-                let base = base.strip_prefix('.').unwrap_or(base);
-                let mut hash = [0u8; 32];
-                if let Some((base, uuid_text)) = base.split_once('-') {
-                    let mut uuid = [0u8; 16];
-                    hex::decode_to_slice(uuid_text, &mut uuid).map_err(|_| ())?;
-                    if ext == "data" {
-                        hex::decode_to_slice(base, &mut hash).map_err(|_| ())?;
-                        Ok(Self::PartialData(hash.into(), uuid))
-                    } else if ext == OUTBOARD_EXT {
-                        hex::decode_to_slice(base, &mut hash).map_err(|_| ())?;
-                        Ok(Self::PartialOutboard(hash.into(), uuid))
-                    } else {
-                        Err(())
-                    }
-                } else if ext == "meta" {
-                    let data = hex::decode(base).map_err(|_| ())?;
-                    Ok(Self::Meta(data))
-                } else {
-                    hex::decode_to_slice(base, &mut hash).map_err(|_| ())?;
-                    if ext == "data" {
-                        Ok(Self::Data(hash.into()))
-                    } else if ext == OUTBOARD_EXT {
-                        Ok(Self::Outboard(hash.into()))
-                    } else if ext == "paths" {
-                        Ok(Self::Paths(hash.into()))
-                    } else {
-                        Err(())
-                    }
-                }
-            }
-        }
-
-        #[derive(Debug, Default)]
-        struct EntryPaths {
-            data: Option<(PathBuf, u64)>,
-            outboard: Option<(PathBuf, u64)>,
-            external: Vec<(PathBuf, u64)>,
-            #[allow(clippy::type_complexity)]
-            partial: BTreeMap<[u8; 16], (Option<(PathBuf, u64)>, Option<(PathBuf, u64)>)>,
-        }
-
-        fn copy_outboard(src: &Path, tgt: &Path) -> io::Result<()> {
-            let mut data = std::fs::read(src)?;
-            if data.len() % 64 != 8 {
-                return Err(io::Error::new(
-                    io::ErrorKind::InvalidData,
-                    "outboard without length prefix",
-                ));
-            }
-            data.splice(0..8, []);
-            std::fs::write(tgt, data)
-        }
-
-        let FlatStorePaths {
-            complete: complete_path,
-            partial: partial_path,
-            meta: meta_path,
-        } = &paths;
-        let mut index = BTreeMap::<Hash, EntryPaths>::new();
-        let mut have_partial = false;
-        let mut have_complete = false;
-        let mut have_meta = false;
-        if partial_path.exists() {
-            tracing::info!("importing partial data from {:?}", partial_path);
-            for entry in std::fs::read_dir(partial_path)? {
-                let entry = entry?;
-                let path = entry.path();
-                if path.is_file() {
-                    let Ok(meta) = entry.metadata() else {
-                        tracing::warn!("unable to open file {}", path.display());
-                        continue;
-                    };
-                    let size = meta.len();
-                    if let Ok(purpose) = FileName::from_path(&path) {
-                        match purpose {
-                            FileName::PartialData(hash, uuid) => {
-                                let m = index.entry(hash).or_default();
-                                m.partial.entry(uuid).or_default().0 = Some((path, size));
-                            }
-                            FileName::PartialOutboard(hash, uuid) => {
-                                let m = index.entry(hash).or_default();
-                                m.partial.entry(uuid).or_default().0 = Some((path, size));
-                            }
-                            _ => {
-                                // silently ignore other files, there could be a valid reason for them
-                            }
-                        }
-                    }
-                }
-            }
-            have_partial = true;
-        }
-
-        if complete_path.exists() {
-            tracing::info!("importing complete data from {:?}", complete_path);
-            for entry in std::fs::read_dir(complete_path)? {
-                let entry = entry?;
-                let path = entry.path();
-                if path.is_file() {
-                    let Ok(meta) = entry.metadata() else {
-                        tracing::warn!("unable to open file {}", path.display());
-                        continue;
-                    };
-                    let size = meta.len();
-                    if let Ok(purpose) = FileName::from_path(&path) {
-                        match purpose {
-                            FileName::Data(hash) => {
-                                let m = index.entry(hash).or_default();
-                                m.data = Some((path, size));
-                            }
-                            FileName::Outboard(hash) => {
-                                let m = index.entry(hash).or_default();
-                                m.outboard = Some((path, size));
-                            }
-                            FileName::Paths(hash) => {
-                                let m = index.entry(hash).or_default();
-                                let paths = std::fs::read(path)?;
-                                let paths: BTreeSet<PathBuf> = postcard::from_bytes(&paths)
-                                    .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
-                                for path in paths {
-                                    let Ok(meta) = path.metadata() else {
-                                        tracing::warn!(
-                                            "unable to open external file {}",
-                                            path.display()
-                                        );
-                                        continue;
-                                    };
-                                    m.external.push((path, meta.len()));
-                                }
-                            }
-                            _ => {
-                                // silently ignore other files, there could be a valid reason for them
-                            }
-                        }
-                    }
-                }
-            }
-            have_complete = true;
-        }
-
-        let tx = self.db.begin_write()?;
-        let mut blobs_table = tx.open_table(BLOBS_TABLE)?;
-        let mut tags_table = tx.open_table(TAGS_TABLE)?;
-        for (hash, entry) in index {
-            if blobs_table.get(hash)?.is_some() {
-                tracing::info!("hash {} already exists in the db", hash.to_hex());
-                continue;
-            }
-            if let Some((data_path, data_size)) = entry.data {
-                let needs_outboard = data_size > IROH_BLOCK_SIZE.bytes() as u64;
-                let outboard_path = if needs_outboard {
-                    let Some((outboard_path, outboard_size)) = entry.outboard else {
-                        tracing::warn!("missing outboard file for {}", hash.to_hex());
-                        continue;
-                    };
-                    if outboard_size != raw_outboard_size(data_size) + 8 {
-                        tracing::warn!("outboard file has wrong size for {}", hash.to_hex());
-                        continue;
-                    }
-                    Some(outboard_path)
-                } else {
-                    None
-                };
-                if let Err(cause) =
-                    std::fs::rename(data_path, self.path_options.owned_data_path(&hash))
-                {
-                    tracing::error!("failed to move data file: {}", cause);
-                    continue;
-                }
-                if let Some(outboard_path) = outboard_path {
-                    if let Err(cause) = copy_outboard(
-                        &outboard_path,
-                        &self.path_options.owned_outboard_path(&hash),
-                    ) {
-                        tracing::error!("failed to move outboard file: {}", cause);
-                        continue;
-                    }
-                }
-                let entry = EntryState::Complete {
-                    data_location: DataLocation::Owned(data_size),
-                    outboard_location: if needs_outboard {
-                        OutboardLocation::Owned
-                    } else {
-                        OutboardLocation::NotNeeded
-                    },
-                };
-                blobs_table.insert(hash, entry)?;
-                continue;
-            }
-            if !entry.external.is_empty() {
-                let sizes = entry.external.iter().map(|x| x.1).collect::<Vec<_>>();
-                if sizes.iter().min() != sizes.iter().max() {
-                    tracing::warn!("external files for {} have different sizes", hash.to_hex());
-                    continue;
-                }
-                let size = sizes[0];
-                let needs_outboard = size > IROH_BLOCK_SIZE.bytes() as u64;
-                let outboard_path = if needs_outboard {
-                    let Some((outboard_path, outboard_size)) = entry.outboard else {
-                        tracing::warn!("missing outboard file for {}", hash.to_hex());
-                        continue;
-                    };
-                    if outboard_size != raw_outboard_size(size) + 8 {
-                        tracing::warn!("outboard file has wrong size for {}", hash.to_hex());
-                        continue;
-                    }
-                    Some(outboard_path)
-                } else {
-                    None
-                };
-                if let Some(outboard_path) = outboard_path {
-                    if let Err(cause) = copy_outboard(
-                        &outboard_path,
-                        &self.path_options.owned_outboard_path(&hash),
-                    ) {
-                        tracing::error!("failed to move outboard file: {}", cause);
-                        continue;
-                    }
-                }
-                let paths = entry
-                    .external
-                    .into_iter()
-                    .map(|(path, _size)| path)
-                    .collect();
-                let entry = EntryState::Complete {
-                    data_location: DataLocation::External(paths, size),
-                    outboard_location: if needs_outboard {
-                        OutboardLocation::Owned
-                    } else {
-                        OutboardLocation::NotNeeded
-                    },
-                };
-                blobs_table.insert(hash, entry)?;
-                continue;
-            }
-            // partial entries that have data
-            let partial_with_data = entry
-                .partial
-                .into_iter()
-                .filter_map(|(_k, (d, o))| d.map(|d| (d, o)));
-            let largest_partial = partial_with_data.max_by_key(|((_, size), _o)| *size);
-            if let Some(((data_path, data_size), outboard)) = largest_partial {
-                let needs_outboard = data_size >= IROH_BLOCK_SIZE.bytes() as u64;
-                let outboard_path = if needs_outboard {
-                    let Some((outboard_path, _)) = outboard else {
-                        tracing::warn!("missing outboard file for {}", hash.to_hex());
-                        continue;
-                    };
-                    Some(outboard_path)
-                } else {
-                    None
-                };
-                if let Err(cause) =
-                    std::fs::rename(data_path, self.path_options.owned_data_path(&hash))
-                {
-                    tracing::error!("failed to move data file: {}", cause);
-                    continue;
-                }
-                if let Some(outboard_path) = outboard_path {
-                    if let Err(cause) = copy_outboard(
-                        &outboard_path,
-                        &self.path_options.owned_outboard_path(&hash),
-                    ) {
-                        tracing::error!("failed to move outboard file: {}", cause);
-                        continue;
-                    }
-                }
-                let entry = EntryState::Partial { size: None };
-                blobs_table.insert(hash, entry)?;
-                continue;
-            }
-        }
-        // import tags, this is pretty straightforward
-        if meta_path.exists() {
-            tracing::info!("importing metadata from {:?}", meta_path);
-            let tags_path = meta_path.join("tags.meta");
-            if tags_path.exists() {
-                let data = std::fs::read(&tags_path)?;
-                #[allow(clippy::mutable_key_type)]
-                let tags: BTreeMap<Tag, HashAndFormat> = postcard::from_bytes(&data)
-                    .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
-                tracing::debug!("loaded tags. {} entries", tags.len());
-                for (tag, content) in tags {
-                    tags_table.insert(tag, content)?;
-                }
-                std::fs::remove_file(tags_path).ok();
-            };
-            have_meta = true;
-        }
-
-        drop(blobs_table);
-        drop(tags_table);
-        tx.commit()?;
-
-        if have_partial {
-            tracing::trace!("removing flat db partial path {:?}", partial_path);
-            if let Err(cause) = std::fs::remove_dir_all(partial_path) {
-                tracing::error!("failed to remove partial path: {}", cause);
-            }
-        }
-        if have_complete {
-            tracing::trace!("removing flat db complete path {:?}", complete_path);
-            if let Err(cause) = std::fs::remove_dir_all(complete_path) {
-                tracing::error!("failed to remove complete path: {}", cause);
-            }
-        }
-        if have_meta {
-            tracing::trace!("removing flat db meta path {:?}", meta_path);
-            if let Err(cause) = std::fs::remove_dir_all(meta_path) {
-                tracing::error!("failed to remove meta path: {}", cause);
-            }
-        }
-        Ok(have_partial || have_complete)
     }
 
     fn delete(&mut self, hashes: Vec<Hash>) -> ActorResult<()> {
@@ -2658,352 +2319,6 @@ impl Actor {
         }
     }
 
-    fn validate(
-        &mut self,
-        progress: tokio::sync::mpsc::Sender<ValidateProgress>,
-    ) -> ActorResult<()> {
-        let tx = self.db.begin_read()?;
-        let blobs = tx.open_table(BLOBS_TABLE)?;
-        let inline_data = tx.open_table(INLINE_DATA_TABLE)?;
-        let inline_outboard = tx.open_table(INLINE_OUTBOARD_TABLE)?;
-        let tags = tx.open_table(TAGS_TABLE)?;
-        macro_rules! send {
-            ($level:expr, $entry:expr, $($arg:tt)*) => {
-                if let Err(_) = progress.blocking_send(ValidateProgress::ConsistencyCheckUpdate { message: format!($($arg)*), level: $level, entry: $entry }) {
-                    return Ok(());
-                }
-            };
-        }
-        macro_rules! trace {
-            ($($arg:tt)*) => {
-                send!(ValidateLevel::Trace, None, $($arg)*)
-            };
-        }
-        macro_rules! info {
-            ($($arg:tt)*) => {
-                send!(ValidateLevel::Info, None, $($arg)*)
-            };
-        }
-        macro_rules! warn {
-            ($($arg:tt)*) => {
-                send!(ValidateLevel::Warn, None, $($arg)*)
-            };
-        }
-        macro_rules! entry_warn {
-            ($hash:expr, $($arg:tt)*) => {
-                send!(ValidateLevel::Warn, Some($hash), $($arg)*)
-            };
-        }
-        macro_rules! entry_info {
-            ($hash:expr, $($arg:tt)*) => {
-                send!(ValidateLevel::Info, Some($hash), $($arg)*)
-            };
-        }
-        macro_rules! error {
-            ($($arg:tt)*) => {
-                send!(ValidateLevel::Error, None, $($arg)*)
-            };
-        }
-        macro_rules! entry_error {
-            ($hash:expr, $($arg:tt)*) => {
-                send!(ValidateLevel::Error, Some($hash), $($arg)*)
-            };
-        }
-        // first, dump the entire data content at trace level
-        trace!("dumping blobs");
-        match blobs.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    match item {
-                        Ok((k, v)) => {
-                            let hash = k.value();
-                            let entry = v.value();
-                            trace!("blob {} -> {:?}", hash.to_hex(), entry);
-                        }
-                        Err(cause) => {
-                            error!("failed to access blob item: {}", cause);
-                        }
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate blobs: {}", cause);
-            }
-        }
-        trace!("dumping inline_data");
-        match inline_data.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    match item {
-                        Ok((k, v)) => {
-                            let hash = k.value();
-                            let data = v.value();
-                            trace!("inline_data {} -> {:?}", hash.to_hex(), data.len());
-                        }
-                        Err(cause) => {
-                            error!("failed to access inline data item: {}", cause);
-                        }
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate inline_data: {}", cause);
-            }
-        }
-        trace!("dumping inline_outboard");
-        match inline_outboard.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    match item {
-                        Ok((k, v)) => {
-                            let hash = k.value();
-                            let data = v.value();
-                            trace!("inline_outboard {} -> {:?}", hash.to_hex(), data.len());
-                        }
-                        Err(cause) => {
-                            error!("failed to access inline outboard item: {}", cause);
-                        }
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate inline_outboard: {}", cause);
-            }
-        }
-        trace!("dumping tags");
-        match tags.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    match item {
-                        Ok((k, v)) => {
-                            let tag = k.value();
-                            let value = v.value();
-                            trace!("tags {} -> {:?}", tag, value);
-                        }
-                        Err(cause) => {
-                            error!("failed to access tag item: {}", cause);
-                        }
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate tags: {}", cause);
-            }
-        }
-
-        // perform consistency check for each entry
-        info!("validating blobs");
-        // set of a all hashes that are referenced by the blobs table
-        let mut entries = BTreeSet::new();
-        match blobs.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    let Ok((hash, entry)) = item else {
-                        error!("failed to access blob item");
-                        continue;
-                    };
-                    let hash = hash.value();
-                    entries.insert(hash);
-                    entry_info!(hash, "validating blob");
-                    let entry = entry.value();
-                    match entry {
-                        EntryState::Complete {
-                            data_location,
-                            outboard_location,
-                        } => {
-                            let data_size = match data_location {
-                                DataLocation::Inline(_) => {
-                                    let Ok(inline_data) = inline_data.get(hash) else {
-                                        entry_error!(hash, "inline data can not be accessed");
-                                        continue;
-                                    };
-                                    let Some(inline_data) = inline_data else {
-                                        entry_error!(hash, "inline data missing");
-                                        continue;
-                                    };
-                                    inline_data.value().len() as u64
-                                }
-                                DataLocation::Owned(size) => {
-                                    let path = self.path_options.owned_data_path(&hash);
-                                    let Ok(metadata) = path.metadata() else {
-                                        entry_error!(hash, "owned data file does not exist");
-                                        continue;
-                                    };
-                                    if metadata.len() != size {
-                                        entry_error!(
-                                            hash,
-                                            "owned data file size mismatch: {}",
-                                            path.display()
-                                        );
-                                        continue;
-                                    }
-                                    size
-                                }
-                                DataLocation::External(paths, size) => {
-                                    for path in paths {
-                                        let Ok(metadata) = path.metadata() else {
-                                            entry_error!(
-                                                hash,
-                                                "external data file does not exist: {}",
-                                                path.display()
-                                            );
-                                            continue;
-                                        };
-                                        if metadata.len() != size {
-                                            entry_error!(
-                                                hash,
-                                                "external data file size mismatch: {}",
-                                                path.display()
-                                            );
-                                            continue;
-                                        }
-                                    }
-                                    size
-                                }
-                            };
-                            match outboard_location {
-                                OutboardLocation::Inline(_) => {
-                                    let Ok(inline_outboard) = inline_outboard.get(hash) else {
-                                        entry_error!(hash, "inline outboard can not be accessed");
-                                        continue;
-                                    };
-                                    let Some(inline_outboard) = inline_outboard else {
-                                        entry_error!(hash, "inline outboard missing");
-                                        continue;
-                                    };
-                                    let outboard_size = inline_outboard.value().len() as u64;
-                                    if outboard_size != raw_outboard_size(data_size) {
-                                        entry_error!(hash, "inline outboard size mismatch");
-                                    }
-                                }
-                                OutboardLocation::Owned => {
-                                    let Ok(metadata) =
-                                        self.path_options.owned_outboard_path(&hash).metadata()
-                                    else {
-                                        entry_error!(hash, "owned outboard file does not exist");
-                                        continue;
-                                    };
-                                    let outboard_size = metadata.len();
-                                    if outboard_size != raw_outboard_size(data_size) {
-                                        entry_error!(hash, "owned outboard size mismatch");
-                                    }
-                                }
-                                OutboardLocation::NotNeeded => {
-                                    if raw_outboard_size(data_size) != 0 {
-                                        entry_error!(
-                                            hash,
-                                            "outboard not needed but data size is not zero"
-                                        );
-                                    }
-                                }
-                            }
-                        }
-                        EntryState::Partial { .. } => {
-                            if !self.path_options.owned_data_path(&hash).exists() {
-                                entry_error!(hash, "persistent partial entry has no data");
-                            }
-                            if !self.path_options.owned_outboard_path(&hash).exists() {
-                                entry_error!(hash, "persistent partial entry has no outboard");
-                            }
-                        }
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate blobs: {}", cause);
-            }
-        };
-        info!("checking for orphaned inline data");
-        match inline_data.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    let Ok((hash, _)) = item else {
-                        error!("failed to access inline data item");
-                        continue;
-                    };
-                    let hash = hash.value();
-                    if !entries.contains(&hash) {
-                        entry_error!(hash, "orphaned inline data");
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate inline_data: {}", cause);
-            }
-        };
-        info!("checking for orphaned inline outboard data");
-        match inline_outboard.iter() {
-            Ok(iter) => {
-                for item in iter {
-                    let Ok((hash, _)) = item else {
-                        error!("failed to access inline outboard item");
-                        continue;
-                    };
-                    let hash = hash.value();
-                    if !entries.contains(&hash) {
-                        entry_error!(hash, "orphaned inline outboard");
-                    }
-                }
-            }
-            Err(cause) => {
-                error!("failed to iterate inline_outboard: {}", cause);
-            }
-        };
-        info!("checking for unexpected or orphaned files");
-        for entry in self.path_options.data_path.read_dir()? {
-            let entry = entry?;
-            let path = entry.path();
-            if !path.is_file() {
-                warn!("unexpected entry in data directory: {}", path.display());
-                continue;
-            }
-            match path.extension().and_then(|x| x.to_str()) {
-                Some("data") => match path.file_stem().and_then(|x| x.to_str()) {
-                    Some(stem) => {
-                        let mut hash = [0u8; 32];
-                        let Ok(_) = hex::decode_to_slice(stem, &mut hash) else {
-                            warn!("unexpected data file in data directory: {}", path.display());
-                            continue;
-                        };
-                        let hash = Hash::from(hash);
-                        if !entries.contains(&hash) {
-                            entry_warn!(hash, "orphaned data file");
-                        }
-                    }
-                    None => {
-                        warn!("unexpected data file in data directory: {}", path.display());
-                    }
-                },
-                Some("obao4") => match path.file_stem().and_then(|x| x.to_str()) {
-                    Some(stem) => {
-                        let mut hash = [0u8; 32];
-                        let Ok(_) = hex::decode_to_slice(stem, &mut hash) else {
-                            warn!(
-                                "unexpected outboard file in data directory: {}",
-                                path.display()
-                            );
-                            continue;
-                        };
-                        let hash = Hash::from(hash);
-                        if !entries.contains(&hash) {
-                            entry_warn!(hash, "orphaned outboard file");
-                        }
-                    }
-                    None => {
-                        warn!(
-                            "unexpected outboard file in data directory: {}",
-                            path.display()
-                        );
-                    }
-                },
-                _ => {
-                    warn!("unexpected file in data directory: {}", path.display());
-                }
-            }
-        }
-        Ok(())
-    }
-
     fn run(mut self) -> ActorResult<()> {
         while let Ok(msg) = self.msgs.recv() {
             match self.handle_one(msg) {
@@ -3026,7 +2341,7 @@ impl Actor {
             ActorMessage::GetOrCreate { hash, tx } => {
                 tx.send(self.get_or_create(hash)?).ok();
             }
-            ActorMessage::ImportEntry {
+            ActorMessage::Import {
                 content,
                 file,
                 data_size,
@@ -3111,6 +2426,7 @@ enum MsgResult {
     Continue,
 }
 
+/// Export a file by copyign out its content to a new location
 fn export_file_copy(
     temp_tag: TempTag,
     path: PathBuf,
@@ -3124,899 +2440,4 @@ fn export_file_copy(
     progress(size)?;
     drop(temp_tag);
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::io::Cursor;
-    // use std::os::unix::fs::PermissionsExt;
-    use bao_tree::ChunkRanges;
-    use iroh_io::AsyncSliceReaderExt;
-    use std::time::Duration;
-
-    use crate::store::bao_file::raw_outboard;
-    use crate::store::bao_file::test_support::{
-        decode_response_into_batch, make_wire_data, random_test_data, simulate_remote, validate,
-    };
-    use crate::store::{Map as _, MapEntryMut, MapMut, Store as _};
-
-    macro_rules! assert_matches {
-        ($expression:expr, $pattern:pat) => {
-            match $expression {
-                $pattern => (),
-                _ => panic!("assertion failed: `(expr matches pattern)` \
-                             expression: `{:?}`, pattern: `{}`", $expression, stringify!($pattern)),
-            }
-        };
-        ($expression:expr, $pattern:pat, $($arg:tt)+) => {
-            match $expression {
-                $pattern => (),
-                _ => panic!("{}: expression: `{:?}`, pattern: `{}`", format_args!($($arg)+), $expression, stringify!($pattern)),
-            }
-        };
-    }
-
-    use super::*;
-
-    /// Helper to simulate a slow request.
-    pub fn to_stream(
-        data: &[u8],
-        mtu: usize,
-        delay: std::time::Duration,
-    ) -> impl Stream<Item = io::Result<Bytes>> + 'static {
-        let parts = data
-            .chunks(mtu)
-            .map(Bytes::copy_from_slice)
-            .collect::<Vec<_>>();
-        futures::stream::iter(parts)
-            .then(move |part| async move {
-                tokio::time::sleep(delay).await;
-                io::Result::Ok(part)
-            })
-            .boxed()
-    }
-
-    async fn create_test_db() -> (tempfile::TempDir, Store) {
-        let _ = tracing_subscriber::fmt::try_init();
-        let testdir = tempfile::tempdir().unwrap();
-        let db_path = testdir.path().join("db.redb");
-        let options = Options {
-            path: PathOptions::new(testdir.path()),
-            inline: Default::default(),
-        };
-        let db = Store::new(db_path, options).await.unwrap();
-        (testdir, db)
-    }
-
-    /// small file that does not have outboard at all
-    const SMALL_SIZE: u64 = 1024;
-    /// medium file that has inline outboard but file data
-    const MID_SIZE: u64 = 1024 * 32;
-    /// large file that has file outboard and file data
-    const LARGE_SIZE: u64 = 1024 * 1024 * 10;
-
-    #[tokio::test]
-    async fn get_cases() {
-        let np = IgnoreProgressSender::<ImportProgress>::default;
-        let (tempdir, db) = create_test_db().await;
-        {
-            let small = Bytes::from(random_test_data(SMALL_SIZE as usize));
-            let (_outboard, hash) = raw_outboard(&small);
-            let res = db.get(&hash).await.unwrap();
-            assert_matches!(res, None);
-            let tt = db
-                .import_bytes(small.clone(), BlobFormat::Raw)
-                .await
-                .unwrap();
-            let res = db.get(&hash).await.unwrap();
-            let Some(entry) = res else {
-                panic!("entry not found");
-            };
-            let actual = entry
-                .data_reader()
-                .await
-                .unwrap()
-                .read_to_end()
-                .await
-                .unwrap();
-            assert_eq!(actual, small);
-            drop(tt);
-        }
-        {
-            let mid = Bytes::from(random_test_data(MID_SIZE as usize));
-            let (_outboard, hash) = raw_outboard(&mid);
-            let res = db.get(&hash).await.unwrap();
-            assert_matches!(res, None);
-            let tt = db.import_bytes(mid.clone(), BlobFormat::Raw).await.unwrap();
-            let res = db.get(&hash).await.unwrap();
-            let Some(entry) = res else {
-                panic!("entry not found");
-            };
-            let actual = entry
-                .data_reader()
-                .await
-                .unwrap()
-                .read_to_end()
-                .await
-                .unwrap();
-            assert_eq!(actual, mid);
-            drop(tt);
-        }
-        {
-            let large = Bytes::from(random_test_data(LARGE_SIZE as usize));
-            let (_outboard, hash) = raw_outboard(&large);
-            let res = db.get(&hash).await.unwrap();
-            assert_matches!(res, None);
-            let tt = db
-                .import_bytes(large.clone(), BlobFormat::Raw)
-                .await
-                .unwrap();
-            let res = db.get(&hash).await.unwrap();
-            let Some(entry) = res else {
-                panic!("entry not found");
-            };
-            let actual = entry
-                .data_reader()
-                .await
-                .unwrap()
-                .read_to_end()
-                .await
-                .unwrap();
-            assert_eq!(actual, large);
-            drop(tt);
-        }
-        {
-            let mid = random_test_data(MID_SIZE as usize);
-            let path = tempdir.path().join("mid.data");
-            std::fs::write(&path, &mid).unwrap();
-            let (_outboard, hash) = raw_outboard(&mid);
-            let res = db.get(&hash).await.unwrap();
-            assert_matches!(res, None);
-            let tt = db
-                .import_file(path, ImportMode::TryReference, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            let res = db.get(&hash).await.unwrap();
-            let Some(entry) = res else {
-                panic!("entry not found");
-            };
-            let actual = entry
-                .data_reader()
-                .await
-                .unwrap()
-                .read_to_end()
-                .await
-                .unwrap();
-            assert_eq!(actual, mid);
-            drop(tt);
-        }
-        drop(tempdir);
-    }
-
-    #[tokio::test]
-    async fn get_or_create_cases() {
-        let (tempdir, db) = create_test_db().await;
-        {
-            const SIZE: u64 = SMALL_SIZE;
-            let data = random_test_data(SIZE as usize);
-            let (hash, reader) = simulate_remote(&data);
-            let entry = db.get_or_create(hash, 0).await.unwrap();
-            {
-                let state = db.entry_state(hash).await.unwrap();
-                assert_eq!(state.db, None);
-                assert_matches!(state.mem, Some(_));
-            }
-            let writer = entry.batch_writer().await.unwrap();
-            decode_response_into_batch(hash, IROH_BLOCK_SIZE, ChunkRanges::all(), reader, writer)
-                .await
-                .unwrap();
-            {
-                let state = db.entry_state(hash).await.unwrap();
-                assert_matches!(state.mem, Some(_));
-                assert_matches!(state.db, None);
-            }
-            db.insert_complete(entry.clone()).await.unwrap();
-            {
-                let state = db.entry_state(hash).await.unwrap();
-                assert_matches!(state.mem, Some(_));
-                assert_matches!(
-                    state.db,
-                    Some(EntryState::Complete {
-                        data_location: DataLocation::Inline(_),
-                        ..
-                    })
-                );
-            }
-            drop(entry);
-            // sync so we know the msg sent on drop is processed
-            db.sync().await.unwrap();
-            let state = db.entry_state(hash).await.unwrap();
-            assert_matches!(state.mem, None);
-        }
-        {
-            const SIZE: u64 = MID_SIZE;
-            let data = random_test_data(SIZE as usize);
-            let (hash, reader) = simulate_remote(&data);
-            let entry = db.get_or_create(hash, 0).await.unwrap();
-            {
-                let state = db.entry_state(hash).await.unwrap();
-                assert_eq!(state.db, None);
-                assert_matches!(state.mem, Some(_));
-            }
-            let writer = entry.batch_writer().await.unwrap();
-            decode_response_into_batch(hash, IROH_BLOCK_SIZE, ChunkRanges::all(), reader, writer)
-                .await
-                .unwrap();
-            {
-                let state = db.entry_state(hash).await.unwrap();
-                assert_matches!(state.mem, Some(_));
-                assert_matches!(state.db, Some(EntryState::Partial { .. }));
-            }
-            db.insert_complete(entry.clone()).await.unwrap();
-            {
-                let state = db.entry_state(hash).await.unwrap();
-                assert_matches!(state.mem, Some(_));
-                assert_matches!(
-                    state.db,
-                    Some(EntryState::Complete {
-                        data_location: DataLocation::Owned(SIZE),
-                        ..
-                    })
-                );
-            }
-            drop(entry);
-            // // sync so we know the msg sent on drop is processed
-            // db.sync().await.unwrap();
-            // let state = db.entry_state(hash).await.unwrap();
-            // assert_matches!(state.mem, None);
-        }
-
-        drop(tempdir);
-    }
-
-    /// Import mem cases, small (data inline, outboard none), mid (data file, outboard inline), large (data file, outboard file)
-    #[tokio::test]
-    async fn import_mem_cases() {
-        let (tempdir, db) = create_test_db().await;
-        {
-            const SIZE: u64 = SMALL_SIZE;
-            let small = Bytes::from(random_test_data(SIZE as usize));
-            let (outboard, hash) = raw_outboard(&small);
-            let tt = db
-                .import_bytes(small.clone(), BlobFormat::Raw)
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::Inline(small),
-                outboard_location: OutboardLocation::NotNeeded,
-            };
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert!(outboard.is_empty());
-        }
-        {
-            const SIZE: u64 = MID_SIZE;
-            let mid = Bytes::from(random_test_data(SIZE as usize));
-            let (outboard, hash) = raw_outboard(&mid);
-            let tt = db.import_bytes(mid.clone(), BlobFormat::Raw).await.unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::<Bytes>::Complete {
-                data_location: DataLocation::Owned(SIZE),
-                outboard_location: OutboardLocation::Inline(Bytes::from(outboard)),
-            };
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(mid, std::fs::read(db.owned_data_path(&hash)).unwrap());
-        }
-        {
-            const SIZE: u64 = LARGE_SIZE;
-            let large = Bytes::from(random_test_data(SIZE as usize));
-            let (outboard, hash) = raw_outboard(&large);
-            let tt = db
-                .import_bytes(large.clone(), BlobFormat::Raw)
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::<Bytes>::Complete {
-                data_location: DataLocation::Owned(SIZE),
-                outboard_location: OutboardLocation::Owned,
-            };
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(large, std::fs::read(db.owned_data_path(&hash)).unwrap());
-            assert_eq!(
-                outboard,
-                std::fs::read(db.owned_outboard_path(&hash)).unwrap()
-            );
-        }
-        drop(tempdir);
-    }
-
-    /// Import mem cases, small (data inline, outboard none), mid (data file, outboard inline), large (data file, outboard file)
-    #[tokio::test]
-    async fn import_stream_cases() {
-        let np = IgnoreProgressSender::<ImportProgress>::default;
-        let (tempdir, db) = create_test_db().await;
-        {
-            const SIZE: u64 = SMALL_SIZE;
-            let small = Bytes::from(random_test_data(SIZE as usize));
-            let (outboard, hash) = raw_outboard(&small);
-            let (tt, size) = db
-                .import_stream(
-                    to_stream(&small, 100, Duration::from_millis(1)),
-                    BlobFormat::Raw,
-                    np(),
-                )
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::Inline(small),
-                outboard_location: OutboardLocation::NotNeeded,
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert!(outboard.is_empty());
-        }
-        {
-            const SIZE: u64 = MID_SIZE;
-            let mid = Bytes::from(random_test_data(SIZE as usize));
-            let (outboard, hash) = raw_outboard(&mid);
-            let (tt, size) = db
-                .import_stream(
-                    to_stream(&mid, 1000, Duration::from_millis(1)),
-                    BlobFormat::Raw,
-                    np(),
-                )
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::<Bytes>::Complete {
-                data_location: DataLocation::Owned(SIZE),
-                outboard_location: OutboardLocation::Inline(Bytes::from(outboard)),
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(mid, std::fs::read(db.owned_data_path(&hash)).unwrap());
-        }
-        {
-            const SIZE: u64 = LARGE_SIZE;
-            let large = Bytes::from(random_test_data(SIZE as usize));
-            let (outboard, hash) = raw_outboard(&large);
-            let (tt, size) = db
-                .import_stream(
-                    to_stream(&large, 100000, Duration::from_millis(1)),
-                    BlobFormat::Raw,
-                    np(),
-                )
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::<Bytes>::Complete {
-                data_location: DataLocation::Owned(SIZE),
-                outboard_location: OutboardLocation::Owned,
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(large, std::fs::read(db.owned_data_path(&hash)).unwrap());
-            assert_eq!(
-                outboard,
-                std::fs::read(db.owned_outboard_path(&hash)).unwrap()
-            );
-        }
-        drop(tempdir);
-    }
-
-    /// Import file cases, small (data inline, outboard none), mid (data file, outboard inline), large (data file, outboard file)
-    #[tokio::test]
-    async fn import_file_cases() {
-        let np = IgnoreProgressSender::<ImportProgress>::default;
-        let (tempdir, db) = create_test_db().await;
-        {
-            const SIZE: u64 = SMALL_SIZE;
-            let small = Bytes::from(random_test_data(SIZE as usize));
-            let path = tempdir.path().join("small.data");
-            std::fs::write(&path, &small).unwrap();
-            let (outboard, hash) = raw_outboard(&small);
-            let (tt, size) = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::Inline(small),
-                outboard_location: OutboardLocation::NotNeeded,
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert!(outboard.is_empty());
-        }
-        {
-            const SIZE: u64 = MID_SIZE;
-            let mid = Bytes::from(random_test_data(SIZE as usize));
-            let path = tempdir.path().join("mid.data");
-            std::fs::write(&path, &mid).unwrap();
-            let (outboard, hash) = raw_outboard(&mid);
-            let (tt, size) = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::Owned(SIZE),
-                outboard_location: OutboardLocation::Inline(Bytes::from(outboard)),
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(mid, std::fs::read(db.owned_data_path(&hash)).unwrap());
-        }
-        {
-            const SIZE: u64 = LARGE_SIZE;
-            let large = Bytes::from(random_test_data(SIZE as usize));
-            let path = tempdir.path().join("mid.data");
-            std::fs::write(&path, &large).unwrap();
-            let (outboard, hash) = raw_outboard(&large);
-            let (tt, size) = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::Owned(SIZE),
-                outboard_location: OutboardLocation::Owned,
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(large, std::fs::read(db.owned_data_path(&hash)).unwrap());
-            assert_eq!(
-                outboard,
-                std::fs::read(db.owned_outboard_path(&hash)).unwrap()
-            );
-        }
-        drop(tempdir);
-    }
-
-    #[tokio::test]
-    async fn import_file_reference_cases() {
-        let np = IgnoreProgressSender::<ImportProgress>::default;
-        let (tempdir, db) = create_test_db().await;
-        {
-            const SIZE: u64 = SMALL_SIZE;
-            let small = Bytes::from(random_test_data(SIZE as usize));
-            let path = tempdir.path().join("small.data");
-            std::fs::write(&path, &small).unwrap();
-            let (outboard, hash) = raw_outboard(&small);
-            let (tt, size) = db
-                .import_file(path, ImportMode::TryReference, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::Inline(small),
-                outboard_location: OutboardLocation::NotNeeded,
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert!(outboard.is_empty());
-        }
-        {
-            const SIZE: u64 = MID_SIZE;
-            let mid = Bytes::from(random_test_data(SIZE as usize));
-            let path = tempdir.path().join("mid.data");
-            std::fs::write(&path, &mid).unwrap();
-            let (outboard, hash) = raw_outboard(&mid);
-            let (tt, size) = db
-                .import_file(
-                    path.clone(),
-                    ImportMode::TryReference,
-                    BlobFormat::Raw,
-                    np(),
-                )
-                .await
-                .unwrap();
-            let actual = db.entry_state(*tt.hash()).await.unwrap();
-            let expected = EntryState::Complete {
-                data_location: DataLocation::External(vec![path.clone()], SIZE),
-                outboard_location: OutboardLocation::Inline(Bytes::from(outboard)),
-            };
-            assert_eq!(size, SIZE);
-            assert_eq!(tt.hash(), &hash);
-            assert_eq!(actual.db, Some(expected));
-            assert_eq!(mid, std::fs::read(path).unwrap());
-            assert!(!db.owned_data_path(&hash).exists());
-        }
-        drop(tempdir);
-    }
-
-    #[tokio::test]
-    async fn import_file_error_cases() {
-        let np = IgnoreProgressSender::<ImportProgress>::default;
-        let (tempdir, db) = create_test_db().await;
-        // relative path is not allowed
-        {
-            let path = PathBuf::from("relativepath.data");
-            let cause = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap_err();
-            assert_eq!(cause.kind(), io::ErrorKind::InvalidInput);
-        }
-        // file does not exist
-        {
-            let path = tempdir.path().join("pathdoesnotexist.data");
-            let cause = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap_err();
-            assert_eq!(cause.kind(), io::ErrorKind::InvalidInput);
-        }
-        // file is a directory
-        {
-            let path = tempdir.path().join("pathisdir.data");
-            std::fs::create_dir_all(&path).unwrap();
-            let cause = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap_err();
-            assert_eq!(cause.kind(), io::ErrorKind::InvalidInput);
-        }
-        // // file is not readable for the store
-        // #[cfg(unix)]
-        // {
-        //     let path = tempdir.path().join("forbidden.data");
-        //     let data = random_test_data(1024);
-        //     std::fs::write(&path, &data).unwrap();
-        //     std::fs::set_permissions(&path, PermissionsExt::from_mode(0o0)).unwrap();
-        //     let cause = db
-        //         .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-        //         .await
-        //         .unwrap_err();
-        //     assert_eq!(cause.kind(), io::ErrorKind::PermissionDenied);
-        // }
-        drop(tempdir);
-    }
-
-    // #[cfg(unix)]
-    // #[tokio::test]
-    // async fn import_file_tempdir_readonly() {
-    //     let np = IgnoreProgressSender::<ImportProgress>::default;
-    //     let (tempdir, db) = create_test_db().await;
-    //     // temp dir is readonly, this is a bit mean since we mess with the internals of the store
-    //     {
-    //         let temp_dir = db.0.temp_file_name().parent().unwrap().to_owned();
-    //         std::fs::set_permissions(temp_dir, PermissionsExt::from_mode(0o0)).unwrap();
-    //         let path = tempdir.path().join("mid.data");
-    //         let data = random_test_data(MID_SIZE as usize);
-    //         std::fs::write(&path, &data).unwrap();
-    //         let cause = db
-    //             .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-    //             .await
-    //             .unwrap_err();
-    //         assert_eq!(cause.kind(), io::ErrorKind::PermissionDenied);
-    //     }
-    //     drop(tempdir);
-    // }
-
-    // #[cfg(unix)]
-    // #[tokio::test]
-    // async fn import_file_datadir_readonly() {
-    //     let np = IgnoreProgressSender::<ImportProgress>::default;
-    //     let (tempdir, db) = create_test_db().await;
-    //     // temp dir is readonly, this is a bit mean since we mess with the internals of the store
-    //     {
-    //         let data_dir = db.0.path_options.data_path.to_owned();
-    //         std::fs::set_permissions(data_dir, PermissionsExt::from_mode(0o0)).unwrap();
-    //         let path = tempdir.path().join("mid.data");
-    //         let data = random_test_data(MID_SIZE as usize);
-    //         std::fs::write(&path, &data).unwrap();
-    //         let cause = db
-    //             .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-    //             .await
-    //             .unwrap_err();
-    //         assert_eq!(cause.kind(), io::ErrorKind::PermissionDenied);
-    //     }
-    //     drop(tempdir);
-    // }
-
-    /// tests that owned wins over external in both cases
-    #[tokio::test]
-    async fn import_file_overwrite() {
-        let np = IgnoreProgressSender::<ImportProgress>::default;
-        let (tempdir, db) = create_test_db().await;
-        // overwrite external with owned
-        {
-            let path = tempdir.path().join("mid.data");
-            let data = random_test_data(MID_SIZE as usize);
-            let (_outboard, hash) = raw_outboard(&data);
-            std::fs::write(&path, &data).unwrap();
-            let (tt1, size1) = db
-                .import_file(path.clone(), ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            assert_eq!(size1, MID_SIZE);
-            assert_eq!(tt1.hash(), &hash);
-            let state = db.entry_state(hash).await.unwrap();
-            assert_matches!(
-                state.db,
-                Some(EntryState::Complete {
-                    data_location: DataLocation::Owned(_),
-                    ..
-                })
-            );
-            let (tt2, size2) = db
-                .import_file(path, ImportMode::TryReference, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            assert_eq!(size2, MID_SIZE);
-            assert_eq!(tt2.hash(), &hash);
-            let state = db.entry_state(hash).await.unwrap();
-            assert_matches!(
-                state.db,
-                Some(EntryState::Complete {
-                    data_location: DataLocation::Owned(_),
-                    ..
-                })
-            );
-        }
-        {
-            let path = tempdir.path().join("mid2.data");
-            let data = random_test_data(MID_SIZE as usize);
-            let (_outboard, hash) = raw_outboard(&data);
-            std::fs::write(&path, &data).unwrap();
-            let (tt1, size1) = db
-                .import_file(
-                    path.clone(),
-                    ImportMode::TryReference,
-                    BlobFormat::Raw,
-                    np(),
-                )
-                .await
-                .unwrap();
-            let state = db.entry_state(hash).await.unwrap();
-            assert_eq!(size1, MID_SIZE);
-            assert_eq!(tt1.hash(), &hash);
-            assert_matches!(
-                state.db,
-                Some(EntryState::Complete {
-                    data_location: DataLocation::External(_, _),
-                    ..
-                })
-            );
-            let (tt2, size2) = db
-                .import_file(path, ImportMode::Copy, BlobFormat::Raw, np())
-                .await
-                .unwrap();
-            let state = db.entry_state(hash).await.unwrap();
-            assert_eq!(size2, MID_SIZE);
-            assert_eq!(tt2.hash(), &hash);
-            assert_matches!(
-                state.db,
-                Some(EntryState::Complete {
-                    data_location: DataLocation::Owned(_),
-                    ..
-                })
-            );
-        }
-        drop(tempdir);
-    }
-
-    /// tests that export works in copy mode
-    #[tokio::test]
-    async fn export_copy_cases() {
-        let np = || Box::new(|_: u64| io::Result::Ok(()));
-        let (tempdir, db) = create_test_db().await;
-        let small = Bytes::from(random_test_data(SMALL_SIZE as usize));
-        let mid = Bytes::from(random_test_data(MID_SIZE as usize));
-        let large = Bytes::from(random_test_data(LARGE_SIZE as usize));
-        let small_tt = db
-            .import_bytes(small.clone(), BlobFormat::Raw)
-            .await
-            .unwrap();
-        let mid_tt = db.import_bytes(mid.clone(), BlobFormat::Raw).await.unwrap();
-        let large_tt = db
-            .import_bytes(large.clone(), BlobFormat::Raw)
-            .await
-            .unwrap();
-        let small_path = tempdir.path().join("small.data");
-        let mid_path = tempdir.path().join("mid.data");
-        let large_path = tempdir.path().join("large.data");
-        db.export(*small_tt.hash(), small_path.clone(), ExportMode::Copy, np())
-            .await
-            .unwrap();
-        assert_eq!(small.to_vec(), std::fs::read(&small_path).unwrap());
-        db.export(*mid_tt.hash(), mid_path.clone(), ExportMode::Copy, np())
-            .await
-            .unwrap();
-        assert_eq!(mid.to_vec(), std::fs::read(&mid_path).unwrap());
-        db.export(*large_tt.hash(), large_path.clone(), ExportMode::Copy, np())
-            .await
-            .unwrap();
-        assert_eq!(large.to_vec(), std::fs::read(&large_path).unwrap());
-        let state = db.entry_state(*small_tt.hash()).await.unwrap();
-        assert_eq!(
-            state.db,
-            Some(EntryState::Complete {
-                data_location: DataLocation::Inline(small),
-                outboard_location: OutboardLocation::NotNeeded,
-            })
-        );
-        let state = db.entry_state(*mid_tt.hash()).await.unwrap();
-        assert_eq!(
-            state.db,
-            Some(EntryState::Complete {
-                data_location: DataLocation::Owned(MID_SIZE),
-                outboard_location: OutboardLocation::Inline(Bytes::from(raw_outboard(&mid).0)),
-            })
-        );
-        let state = db.entry_state(*large_tt.hash()).await.unwrap();
-        assert_eq!(
-            state.db,
-            Some(EntryState::Complete {
-                data_location: DataLocation::Owned(LARGE_SIZE),
-                outboard_location: OutboardLocation::Owned,
-            })
-        );
-    }
-
-    /// tests that export works in reference mode
-    #[tokio::test]
-    async fn export_reference_cases() {
-        let np = || Box::new(|_: u64| io::Result::Ok(()));
-        let (tempdir, db) = create_test_db().await;
-        let small = Bytes::from(random_test_data(SMALL_SIZE as usize));
-        let mid = Bytes::from(random_test_data(MID_SIZE as usize));
-        let large = Bytes::from(random_test_data(LARGE_SIZE as usize));
-        let small_tt = db
-            .import_bytes(small.clone(), BlobFormat::Raw)
-            .await
-            .unwrap();
-        let mid_tt = db.import_bytes(mid.clone(), BlobFormat::Raw).await.unwrap();
-        let large_tt = db
-            .import_bytes(large.clone(), BlobFormat::Raw)
-            .await
-            .unwrap();
-        let small_path = tempdir.path().join("small.data");
-        let mid_path = tempdir.path().join("mid.data");
-        let large_path = tempdir.path().join("large.data");
-        db.export(
-            *small_tt.hash(),
-            small_path.clone(),
-            ExportMode::TryReference,
-            np(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(small.to_vec(), std::fs::read(&small_path).unwrap());
-        db.export(
-            *mid_tt.hash(),
-            mid_path.clone(),
-            ExportMode::TryReference,
-            np(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(mid.to_vec(), std::fs::read(&mid_path).unwrap());
-        db.export(
-            *large_tt.hash(),
-            large_path.clone(),
-            ExportMode::TryReference,
-            np(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(large.to_vec(), std::fs::read(&large_path).unwrap());
-        let state = db.entry_state(*small_tt.hash()).await.unwrap();
-        // small entries will never use external references
-        assert_eq!(
-            state.db,
-            Some(EntryState::Complete {
-                data_location: DataLocation::Inline(small),
-                outboard_location: OutboardLocation::NotNeeded,
-            })
-        );
-        // mid entries should now use external references
-        let state = db.entry_state(*mid_tt.hash()).await.unwrap();
-        assert_eq!(
-            state.db,
-            Some(EntryState::Complete {
-                data_location: DataLocation::External(vec![mid_path], MID_SIZE),
-                outboard_location: OutboardLocation::Inline(Bytes::from(raw_outboard(&mid).0)),
-            })
-        );
-        // large entries should now use external references
-        let state = db.entry_state(*large_tt.hash()).await.unwrap();
-        assert_eq!(
-            state.db,
-            Some(EntryState::Complete {
-                data_location: DataLocation::External(vec![large_path], LARGE_SIZE),
-                outboard_location: OutboardLocation::Owned,
-            })
-        );
-    }
-
-    /// tests that export to a non existing directory fails
-    // #[cfg(unix)]
-    // #[tokio::test]
-    // async fn export_nonexistent_dir() {
-    //     let (tempdir, db) = create_test_db().await;
-    //     let small = Bytes::from(random_test_data(SMALL_SIZE as usize));
-    //     let target_dir = tempdir.path().join("export");
-    //     let small_tt = db
-    //         .import_bytes(small.clone(), BlobFormat::Raw)
-    //         .await
-    //         .unwrap();
-    //     let small_path = target_dir.join("small.data");
-    //     let cause = db
-    //         .export(
-    //             *small_tt.hash(),
-    //             small_path.clone(),
-    //             ExportMode::Copy,
-    //             Box::new(|_: u64| io::Result::Ok(())),
-    //         )
-    //         .await
-    //         .unwrap_err();
-    //     assert_eq!(cause.kind(), io::ErrorKind::PermissionDenied);
-    // }
-
-    #[tokio::test]
-    async fn entry_drop() {
-        let _ = tracing_subscriber::fmt::try_init();
-        let testdir = tempfile::tempdir().unwrap();
-        let db_path = testdir.path().join("test.redb");
-        let options = Options {
-            path: PathOptions::new(testdir.path()),
-            inline: Default::default(),
-        };
-        let db = Store::new(db_path, options).await.unwrap();
-        let data = random_test_data(1024 * 1024);
-        let (_outboard, hash) = raw_outboard(data.as_slice());
-        let entry = db.get_or_create(hash, 0).await.unwrap();
-        let id = entry.0.id;
-        let e2 = entry.clone();
-        assert_eq!(id, e2.0.id);
-        drop(entry);
-        drop(e2);
-        tokio::time::sleep(Duration::from_millis(50)).await;
-        let entry = db.get_or_create(hash, 0).await.unwrap();
-        assert_ne!(id, entry.0.id);
-        drop(db);
-    }
-
-    #[tokio::test]
-    async fn actor_store_smoke() {
-        let testdir = tempfile::tempdir().unwrap();
-        let db_path = testdir.path().join("test.redb");
-        let options = Options {
-            path: PathOptions::new(testdir.path()),
-            inline: Default::default(),
-        };
-        let db = Store::new(db_path, options).await.unwrap();
-        db.dump().await.unwrap();
-        let data = random_test_data(1024 * 1024);
-        #[allow(clippy::single_range_in_vec_init)]
-        let ranges = [0..data.len() as u64];
-        let (hash, chunk_ranges, wire_data) = make_wire_data(&data, &ranges);
-        let handle = db.get_or_create(hash, 0).await.unwrap();
-        decode_response_into_batch(
-            hash,
-            IROH_BLOCK_SIZE,
-            chunk_ranges.clone(),
-            Cursor::new(wire_data),
-            handle.batch_writer().await.unwrap(),
-        )
-        .await
-        .unwrap();
-        validate(&handle.0, &data, &ranges).await;
-        db.insert_complete(handle).await.unwrap();
-        db.sync().await.unwrap();
-        db.dump().await.unwrap();
-    }
 }
